Database access outline
Basic form of the functions for accessing the database have been outlined. Some psuedocode, and some helpers have been added, but all of the LINQ and SQL are still needed. Some sizes have been increased in the database, since the increased size should have negligible impacts on speed, and benefit the functionality.