ImageViewer MVVM
MVVM strating to be set up for the image viewer. I was attempting to allow the image viewer to open both images and video by having swappable user controls, but they have some problems, and it would probably be best to have seperate design for each anyway, so I will probably scrap a lot of this tomorrow. The MVVM set up seems to be working though, so its a good start