Hyperlink
Turns out they all work. The databinding was broken. Apparently the URLs cannot be databound at all. Will have to break MVVM to progress in a timely matter, but that's fine. I want to follow MVVM closely, but breaking it to make something radically easier is acceptable. Keeping both Hyperlink methods in case it can come in use later